Consolidate "math" helpers
