+++ ethereum/spec/src/test/java/tech/pegasys/teku/spec/logic/common/helpers/MathHelpersTest.java
@@ -11,50 +11,71 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.spec.logic.common.util;
+package tech.pegasys.teku.spec.logic.common.helpers;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 
 import org.apache.tuweni.bytes.Bytes;
 import org.apache.tuweni.bytes.Bytes32;
 import org.junit.jupiter.api.Test;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
 
-public class ByteUtilsTest {
+public class MathHelpersTest {
 
   @Test
   void intToBytes32UInt64() {
     assertEquals(
         Bytes32.fromHexString("0x0000000000000000000000000000000000000000000000000000000000000000"),
-        ByteUtils.uintToBytes32(UInt64.ZERO));
+        MathHelpers.uintToBytes32(UInt64.ZERO));
     assertEquals(
         Bytes32.fromHexString("0x0100000000000000000000000000000000000000000000000000000000000000"),
-        ByteUtils.uintToBytes32(UInt64.ONE));
+        MathHelpers.uintToBytes32(UInt64.ONE));
     assertEquals(
         Bytes32.fromHexString("0xffffffffffffffff000000000000000000000000000000000000000000000000"),
-        ByteUtils.uintToBytes32(UInt64.MAX_VALUE));
+        MathHelpers.uintToBytes32(UInt64.MAX_VALUE));
     assertEquals(
         Bytes32.fromHexString("0xefcdab8967452301000000000000000000000000000000000000000000000000"),
-        ByteUtils.uintToBytes32(UInt64.valueOf(0x0123456789abcdefL)));
+        MathHelpers.uintToBytes32(UInt64.valueOf(0x0123456789abcdefL)));
   }
 
   @Test
   void bytesToInt() {
-    assertEquals(UInt64.valueOf(0), ByteUtils.bytesToUInt64(Bytes.fromHexString("0x00")));
-    assertEquals(UInt64.valueOf(1), ByteUtils.bytesToUInt64(Bytes.fromHexString("0x01")));
+    assertEquals(UInt64.valueOf(0), MathHelpers.bytesToUInt64(Bytes.fromHexString("0x00")));
+    assertEquals(UInt64.valueOf(1), MathHelpers.bytesToUInt64(Bytes.fromHexString("0x01")));
     assertEquals(
-        UInt64.valueOf(1), ByteUtils.bytesToUInt64(Bytes.fromHexString("0x0100000000000000")));
+        UInt64.valueOf(1), MathHelpers.bytesToUInt64(Bytes.fromHexString("0x0100000000000000")));
     assertEquals(
         UInt64.valueOf(0x123456789abcdef0L),
-        ByteUtils.bytesToUInt64(Bytes.fromHexString("0xf0debc9a78563412")));
+        MathHelpers.bytesToUInt64(Bytes.fromHexString("0xf0debc9a78563412")));
     assertEquals(
         UInt64.fromLongBits(0xffffffffffffffffL),
-        ByteUtils.bytesToUInt64(Bytes.fromHexString("0xffffffffffffffff")));
+        MathHelpers.bytesToUInt64(Bytes.fromHexString("0xffffffffffffffff")));
     assertEquals(
         UInt64.fromLongBits(0x0000000000000080L),
-        ByteUtils.bytesToUInt64(Bytes.fromHexString("0x8000000000000000")));
+        MathHelpers.bytesToUInt64(Bytes.fromHexString("0x8000000000000000")));
     assertEquals(
         UInt64.fromLongBits(0xffffffffffffff7fL),
-        ByteUtils.bytesToUInt64(Bytes.fromHexString("0x7fffffffffffffff")));
+        MathHelpers.bytesToUInt64(Bytes.fromHexString("0x7fffffffffffffff")));
+  }
+
+  @Test
+  void sqrtOfSquareNumber() {
+    UInt64 actual = MathHelpers.integerSquareRoot(UInt64.valueOf(3481L));
+    UInt64 expected = UInt64.valueOf(59L);
+    assertEquals(expected, actual);
+  }
+
+  @Test
+  void sqrtOfANonSquareNumber() {
+    UInt64 actual = MathHelpers.integerSquareRoot(UInt64.valueOf(27L));
+    UInt64 expected = UInt64.valueOf(5L);
+    assertEquals(expected, actual);
+  }
+
+  @Test
+  void sqrtOfANegativeNumber() {
+    assertThrows(
+        IllegalArgumentException.class, () -> MathHelpers.integerSquareRoot(UInt64.valueOf(-1L)));
   }
 }
